+++ src/DateTime/Clock/RealtimeClock.php
@@ -6,6 +6,10 @@
 
 namespace Ingenerator\PHPUtils\DateTime\Clock;
 
+use DateInterval;
+use DateTimeImmutable;
+use Ingenerator\PHPUtils\DateTime\DateTimeImmutableFactory;
+
 /**
  * Simple wrapper around current date/time methods to allow easy injection of fake time in
  * dependent classes
@@ -14,13 +18,12 @@ namespace Ingenerator\PHPUtils\DateTime\Clock;
  */
 class RealtimeClock
 {
-
     /**
-     * @return \DateTimeImmutable
+     * @return DateTimeImmutable
      */
     public function getDateTime()
     {
-        return new \DateTimeImmutable;
+        return new DateTimeImmutable;
     }
 
     /**
@@ -37,5 +40,31 @@ class RealtimeClock
     public function usleep($microseconds)
     {
         \usleep($microseconds);
+    }
+
+    /**
+     * Calculate a relative date in the past, optionally truncating time to 0 - sugar for getDateTime()->sub()
+     */
+    public function ago(DateInterval $interval, bool $date_only = FALSE): DateTimeImmutable
+    {
+        $result = $this->getDateTime()->sub($interval);
+
+        return match ($date_only) {
+            FALSE => $result,
+            TRUE => DateTimeImmutableFactory::zeroTime($result)
+        };
+    }
+
+    /**
+     * Calculate a relative date in the future, optionally truncating time to 0 - sugar for getDateTime()->add()
+     */
+    public function future(DateInterval $interval, bool $date_only = FALSE): DateTimeImmutable
+    {
+        $result = $this->getDateTime()->add($interval);
+
+        return match ($date_only) {
+            FALSE => $result,
+            TRUE => DateTimeImmutableFactory::zeroTime($result)
+        };
     }
 }